3D model/requirements.txt
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7.1
dash==3.1.1
decorator==5.2.1
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.1.1
fsspec==2025.5.1
idna==3.10
importlib_metadata==8.7.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==1.46.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
open3d==0.19.0
opencv-python==4.12.0.88
packaging==25.0
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
prompt_toolkit==3.0.51
pure_eval==0.2.3
Pygments==2.19.2
pywin32==310
referencing==0.36.2
requests==2.32.4
retrying==1.4.0
rpds-py==0.26.0
setuptools==80.9.0
stack-data==0.6.3
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
traitlets==5.14.3
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.14
zipp==3.23.0
